+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Unreleased
 ----------
-- Removed `UiEvent::Directed` & `UiEvent::Returnable` variants
+- Removed custom event variants & infrastructure
 
 
 0.6.0-alpha.2